Move encryption out of CardComponentStateExt
To ensure we don't trigger onSubmit when encryption fails
