+++ card/src/main/java/com/adyen/checkout/card/internal/ui/state/CardComponentStateExt.kt
@@ -10,7 +10,6 @@
 
 package com.adyen.checkout.card.internal.ui.state
 
-import com.adyen.checkout.card.internal.helper.ExpiryDateParser
 import com.adyen.checkout.card.internal.ui.model.CardComponentParams
 import com.adyen.checkout.card.internal.ui.model.InstallmentModel
 import com.adyen.checkout.card.internal.ui.model.InstallmentPlan
@@ -22,47 +21,18 @@ import com.adyen.checkout.core.components.data.PaymentComponentData
 import com.adyen.checkout.core.components.internal.data.provider.SdkDataProvider
 import com.adyen.checkout.core.components.internal.ui.state.model.getPaymentDataValue
 import com.adyen.checkout.core.components.paymentmethod.CardDetails
-import com.adyen.checkout.core.error.internal.GenericError
-import com.adyen.checkout.core.error.internal.InternalCheckoutError
 import com.adyen.checkout.cse.EncryptedCard
-import com.adyen.checkout.cse.EncryptionException
-import com.adyen.checkout.cse.UnencryptedCard
-import com.adyen.checkout.cse.internal.BaseCardEncryptor
-import com.adyen.checkout.cse.internal.BaseGenericEncryptor
 import com.adyen.threeds2.ThreeDS2Service
 
 @Suppress("ReturnCount", "LongParameterList")
 internal fun CardComponentState.toPaymentComponentState(
-    publicKey: String?,
     componentParams: CardComponentParams,
-    cardEncryptor: BaseCardEncryptor,
-    genericEncryptor: BaseGenericEncryptor,
     sdkDataProvider: SdkDataProvider,
     paymentMethodType: String,
-    onEncryptionFailed: (EncryptionException) -> Unit,
-    onPublicKeyNotFound: (InternalCheckoutError) -> Unit,
     fundingSource: String?,
+    encryptedCard: EncryptedCard,
+    encryptedKcpCardPassword: String?,
 ): CardPaymentComponentState {
-    publicKey ?: run {
-        onPublicKeyNotFound(GenericError("Public key is missing."))
-        return invalidCardPaymentComponentState()
-    }
-
-    val encryptedCard = encryptCard(
-        cardEncryptor = cardEncryptor,
-        publicKey = publicKey,
-        onEncryptionFailed = onEncryptionFailed,
-    ) ?: return invalidCardPaymentComponentState()
-
-    val encryptedKcpCardPassword = kcpCardPassword.getPaymentDataValue()?.let { kcpCardPassword ->
-        encryptKcpCardPassword(
-            genericEncryptor = genericEncryptor,
-            publicKey = publicKey,
-            kcpCardPassword = kcpCardPassword,
-            onEncryptionFailed = onEncryptionFailed,
-        ) ?: return invalidCardPaymentComponentState()
-    }
-
     val cardDetails = createCardDetails(
         encryptedCard = encryptedCard,
         holderName = holderName.getPaymentDataValue(),
@@ -83,55 +53,6 @@ internal fun CardComponentState.toPaymentComponentState(
     )
 
     return createPaymentComponentState(paymentComponentData)
-}
-
-private fun CardComponentState.encryptCard(
-    cardEncryptor: BaseCardEncryptor,
-    publicKey: String,
-    onEncryptionFailed: (EncryptionException) -> Unit,
-): EncryptedCard? {
-    return try {
-        val unencryptedCardBuilder = UnencryptedCard.Builder()
-        cardNumber.getPaymentDataValue()?.let {
-            unencryptedCardBuilder.setNumber(it)
-        }
-
-        securityCode.getPaymentDataValue()?.let {
-            unencryptedCardBuilder.setCvc(it)
-        }
-
-        expiryDate.getPaymentDataValue()?.let {
-            ExpiryDateParser.parseToMonthAndYear(it, returnFullYear = true)
-        }?.let { (expiryMonth, expiryYear) ->
-            unencryptedCardBuilder.setExpiryDate(
-                expiryMonth = expiryMonth,
-                expiryYear = expiryYear,
-            )
-        }
-
-        cardEncryptor.encryptFields(unencryptedCardBuilder.build(), publicKey)
-    } catch (e: EncryptionException) {
-        onEncryptionFailed(e)
-        null
-    }
-}
-
-private fun encryptKcpCardPassword(
-    genericEncryptor: BaseGenericEncryptor,
-    publicKey: String,
-    kcpCardPassword: String,
-    onEncryptionFailed: (EncryptionException) -> Unit,
-): String? {
-    return try {
-        genericEncryptor.encryptField(
-            fieldKeyToEncrypt = ENCRYPTION_KEY_FOR_KCP_PASSWORD,
-            fieldValueToEncrypt = kcpCardPassword,
-            publicKey = publicKey,
-        )
-    } catch (e: EncryptionException) {
-        onEncryptionFailed(e)
-        null
-    }
 }
 
 @Suppress("LongParameterList")
@@ -212,11 +133,6 @@ private fun createPaymentComponentState(
     )
 }
 
-private fun invalidCardPaymentComponentState() = CardPaymentComponentState(
-    data = PaymentComponentData(null, null, null),
-    isValid = false,
-)
-
 /**
  * We only prefill the card brand in the onSubmit payload if we are sure the card has a single brand identified on the
  * backend (reliable) or if the shopper manually selected that brand. In all other cases we cannot reliably assume
@@ -240,5 +156,3 @@ private fun CardComponentState.cardBrand(): CardBrand? {
 private fun CardComponentState.storePaymentMethod(componentParams: CardComponentParams): Boolean? {
     return storePaymentMethod.takeIf { componentParams.showStorePaymentMethod }
 }
-
-private const val ENCRYPTION_KEY_FOR_KCP_PASSWORD = "password"
